+++ webpack.start.js
@@ -21,7 +21,6 @@ module.exports = {
   },
   entry: {
     'index.js': './demo/index.js',
-    'index-preview.js': './demo/index-preview.js',
     'jest-puppeteer.js': './demo/jest-puppeteer.js',
   },
   resolve: {
@@ -112,11 +111,6 @@ module.exports = {
       chunks: ['index.js'],
       filename: './index.html',
       template: './demo/index.html',
-    }),
-    new HtmlWebpackPlugin({
-      chunks: ['index-preview.js'],
-      filename: './index-preview.html',
-      template: './demo/index-preview.html',
     }),
     new HtmlWebpackPlugin({
       chunks: ['jest-puppeteer.js'],